✨💄 Implement JapanMap component
