+++ src/app/[locale]/map/map-page-client.tsx
@@ -1,45 +1,39 @@
 'use client'
 
-import { JapanMap } from '@/app/ui/components/organisms/japan-map/japan-map'
-import { RegionArticleList } from '@/app/ui/components/organisms/region-article-list/region-article-list'
+import { JapanMap } from '@/app/ui/components/organisms/japan-map'
+import { RegionArticleList } from '@/app/ui/components/organisms/region-article-list'
 import type { BlogPostWithHref } from '@/types/blog-post'
+import type { Prefecture } from '@/types/region'
 import { type FC, useState } from 'react'
-import type { Prefecture, Region } from './page'
 
 type Props = {
-  initialRegions: Region[]
   initialAllPosts: BlogPostWithHref[]
   initialPrefectures: Prefecture[]
 }
 
-export const MapPageClient: FC<Props> = ({ initialRegions, initialPrefectures, initialAllPosts }) => {
-  const [regions] = useState<Region[]>(initialRegions)
+export const MapPageClient: FC<Props> = ({ initialPrefectures, initialAllPosts }) => {
   const [prefectures] = useState<Prefecture[]>(initialPrefectures)
   const [allPosts] = useState<BlogPostWithHref[]>(initialAllPosts)
   const [selectedRegionId, setSelectedRegionId] = useState<string | null>(null)
 
-  // Filtering logic remains the same
   const filteredPosts = selectedRegionId ? allPosts.filter((post) => post.contentfulMetadata?.concepts?.some((c) => c?.id === selectedRegionId)) : allPosts
 
-  // Handler for region selection remains the same
   const handleRegionSelect = (regionId: string) => {
-    setSelectedRegionId(regionId === selectedRegionId ? null : regionId) // Toggle selection
+    setSelectedRegionId(regionId === selectedRegionId ? null : regionId)
   }
-
-  // No useEffect needed here for initial data fetching as it's passed via props
 
   return (
     <div className="container mx-auto px-4 py-8">
-      <h1 className="text-3xl font-bold mb-6">Explore Articles by Region</h1>
+      <h1 className="text-3xl font-bold mb-6">Explore Articles by Prefecture</h1>
       <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
         <div className="md:col-span-2">
-          <JapanMap regions={regions} prefectures={prefectures} onSelectRegion={handleRegionSelect} selectedRegionId={selectedRegionId} />
+          <JapanMap prefectures={prefectures} onSelectRegion={handleRegionSelect} selectedRegionId={selectedRegionId} />
         </div>
         <div>
           <h2 className="text-2xl font-semibold mb-4">
-            {selectedRegionId ? `Articles in ${regions.find((r) => r.id === selectedRegionId)?.name || 'Selected Region'}` : 'All Articles'}
+            {selectedRegionId ? `Articles in ${prefectures.find((r) => r.id === selectedRegionId)?.name || 'Selected Region'}` : 'All Articles'}
           </h2>
-          <RegionArticleList articles={filteredPosts} />
+          <RegionArticleList blogPosts={filteredPosts} />
         </div>
       </div>
     </div>
