+++ src/app/ui/colors.ts
@@ -6,6 +6,16 @@ export enum COLORS {
   WHITE = '#FFFFFF',
   TRANSPARENT = 'transparent',
 
+  PRIMARY_50 = '#FBEDED',
+  PRIMARY_100 = '#F8D0CD',
+  PRIMARY_200 = '#F0A19C',
+  PRIMARY_300 = '#E8726A',
+  PRIMARY_400 = '#DB4339',
+  PRIMARY_600 = '#A82219',
+  PRIMARY_700 = '#8D1B15',
+  PRIMARY_800 = '#721510',
+  PRIMARY_900 = '#57100C',
+
   SLATE_50 = '#F8FAFC',
   SLATE_100 = '#F1F5F9',
   SLATE_200 = '#E2E8F0',
